face/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
cffi==1.17.1
click==8.1.8
cryptography==44.0.2
dlib==19.24.6
dnspython==2.7.0
face-recognition==1.3.0
fastapi==0.115.11
h11==0.14.0
idna==3.10
mongoengine==0.29.1
numpy==2.2.3
opencv-python==4.11.0.86
pillow==11.1.0
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pymongo==4.11.1
python-dotenv==1.0.1
sniffio==1.3.1
SQLAlchemy==2.0.38
starlette==0.46.0
typing_extensions==4.12.2
uvicorn==0.34.0
